STYLE: removed some double semicola (semicolons, semicoli?) FIX: missing break statements in switches
git-svn-id: https://svn.rwth-aachen.de/repos/vrgroup-svn/projects/Vista/trunk@38991 33c38b33-09d2-4561-8d4d-1bfa2f58f885

+++ VistaCoreLibs/VistaDataFlowNet/VdfnHistoryProjectNode.cpp
@@ -110,7 +110,7 @@ bool VdfnHistoryProjectNode::DoEvalNode()
 
 			m_pIterateState->SetValue( m_nUnprocessedMeasures, GetUpdateTimeStamp() );
 
-			if( --m_nUnprocessedMeasures != ~0 )
+			if( --m_nUnprocessedMeasures != (unsigned int)~0 )
 				pMeasure = m_pHistory->GetValue()->m_oHistory.GetPast( m_nUnprocessedMeasures );
 			else
 			{
